// apps/mobile/http/generated/models/PostOrganizationsSlugEmployees.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type PostOrganizationsSlugEmployeesPathParams = {
  /**
   * @type string
   */
  slug: string;
};

/**
 * @description Default Response
 */
export type PostOrganizationsSlugEmployees201 = {
  /**
   * @type string, uuid
   */
  employeeId: string;
};

export const postOrganizationsSlugEmployeesMutationRequestPixKeyTypeEnum = {
  CPF: "CPF",
  CNPJ: "CNPJ",
  EMAIL: "EMAIL",
  PHONE: "PHONE",
  RANDOM: "RANDOM",
} as const;

export type PostOrganizationsSlugEmployeesMutationRequestPixKeyTypeEnumKey =
  (typeof postOrganizationsSlugEmployeesMutationRequestPixKeyTypeEnum)[keyof typeof postOrganizationsSlugEmployeesMutationRequestPixKeyTypeEnum];

export type PostOrganizationsSlugEmployeesMutationRequest = {
  /**
   * @type string
   */
  name: string;
  /**
   * @type string | undefined
   */
  role?: string;
  /**
   * @type string
   */
  email: string;
  /**
   * @type string | undefined
   */
  phone?: string;
  /**
   * @type string | undefined
   */
  department?: string;
  /**
   * @type string | undefined
   */
  cpf?: string;
  /**
   * @type string | undefined
   */
  pixKeyType?: PostOrganizationsSlugEmployeesMutationRequestPixKeyTypeEnumKey;
  /**
   * @type string | undefined
   */
  pixKey?: string;
  /**
   * @type string | undefined
   */
  paymentNotes?: string;
  /**
   * @type string | undefined
   */
  country?: string;
  /**
   * @type string | undefined
   */
  state?: string;
  /**
   * @type string | undefined
   */
  city?: string;
  /**
   * @type string | undefined
   */
  street?: string;
  /**
   * @type string | undefined
   */
  zipCode?: string;
  /**
   * @type string | undefined
   */
  neighborhood?: string;
  /**
   * @type string | undefined
   */
  number?: string;
  /**
   * @type string | undefined
   */
  complement?: string;
  /**
   * @type string, uuid
   */
  companyId: string;
  /**
   * @type string | undefined, uuid
   */
  unitId?: string;
};

export type PostOrganizationsSlugEmployeesMutationResponse =
  PostOrganizationsSlugEmployees201;

export type PostOrganizationsSlugEmployeesMutation = {
  Response: PostOrganizationsSlugEmployees201;
  Request: PostOrganizationsSlugEmployeesMutationRequest;
  PathParams: PostOrganizationsSlugEmployeesPathParams;
  Errors: any;
};